Adjust the cli directory structure.Fixed some bug.
