+++ cmd/osdsctl/main.go
@@ -24,12 +24,19 @@ import (
 	"log"
 	"os"
 
-	"github.com/opensds/opensds/pkg/cli"
+	"github.com/opensds/opensds/osdsctl"
+	"github.com/opensds/opensds/pkg/utils"
 )
 
+const logDir = "/var/log/opensds"
+
 func main() {
+	// Create the default log directory
+	if exist, _ := utils.PathExists(logDir); !exist {
+		os.MkdirAll(logDir, 0755)
+	}
 	// Open OpenSDS CLI service log file
-	f, err := os.OpenFile("/var/log/opensds/osdsctl.log", os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
+	f, err := os.OpenFile(logDir+"/osdsctl.log", os.O_RDWR|os.O_CREATE|os.O_APPEND, 0666)
 	if err != nil {
 		fmt.Errorf("Error opening file:%v", err)
 		os.Exit(1)
