+++ osdsctl/dock.go
@@ -11,6 +11,7 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 /*
 This module implements a entry into the OpenSDS service.
 
@@ -25,62 +26,61 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var poolCommand = &cobra.Command{
-	Use:   "pool",
-	Short: "manage OpenSDS pool resources",
-	Run:   poolAction,
+var dockCommand = &cobra.Command{
+	Use:   "dock",
+	Short: "manage OpenSDS dock resources",
+	Run:   dockAction,
 }
 
-var poolShowCommand = &cobra.Command{
-	Use:   "show <pool id>",
-	Short: "show information of specified pool",
-	Run:   poolShowAction,
+var dockShowCommand = &cobra.Command{
+	Use:   "show <dock id>",
+	Short: "show information of specified dock",
+	Run:   dockShowAction,
 }
 
-var poolListCommand = &cobra.Command{
+var dockListCommand = &cobra.Command{
 	Use:   "list",
-	Short: "get all pool resources",
-	Run:   poolListAction,
+	Short: "get all dock resources",
+	Run:   dockListAction,
 }
 
 func init() {
-	poolCommand.AddCommand(poolShowCommand)
-	poolCommand.AddCommand(poolListCommand)
+	dockCommand.AddCommand(dockShowCommand)
+	dockCommand.AddCommand(dockListCommand)
 }
 
-func poolAction(cmd *cobra.Command, args []string) {
+func dockAction(cmd *cobra.Command, args []string) {
 	cmd.Usage()
 	os.Exit(1)
 }
 
-func poolShowAction(cmd *cobra.Command, args []string) {
+func dockShowAction(cmd *cobra.Command, args []string) {
 	if len(args) != 1 {
 		fmt.Println("The number of args is not correct!")
 		cmd.Usage()
 		os.Exit(1)
 	}
-	pols, err := client.GetPool(args[0])
+	resp, err := client.GetDock(args[0])
 	if err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
-	keys := KeyList{"Id", "CreatedAt", "UpdatedAt", "Name", "Description", "Status", "DockId",
-		"AvailabilityZone", "TotalCapacity", "FreeCapacity", "StorageType", "Parameters"}
-	PrintDict(pols, keys, FormatterList{})
+	keys := KeyList{"Id", "CreatedAt", "UpdatedAt", "Name", "Description", "Endpoint", "DriverName", "Parameters"}
+	PrintDict(resp, keys, FormatterList{})
 }
 
-func poolListAction(cmd *cobra.Command, args []string) {
+func dockListAction(cmd *cobra.Command, args []string) {
 	if len(args) != 0 {
 		fmt.Println("The number of args is not correct!")
 		cmd.Usage()
 		os.Exit(1)
 	}
-	pols, err := client.ListPools()
+
+	resp, err := client.ListDocks()
 	if err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
-	keys := KeyList{"Id", "Name", "Description", "Status", "AvailabilityZone", "TotalCapacity", "FreeCapacity"}
-	PrintList(pols, keys, FormatterList{})
+	keys := KeyList{"Id", "Name", "Description", "Endpoint", "DriverName", "Parameters"}
+	PrintList(resp, keys, FormatterList{})
 }
-
