+++ osdsctl/pool.go
@@ -11,7 +11,6 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
 /*
 This module implements a entry into the OpenSDS service.
 
@@ -26,62 +25,61 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var dockCommand = &cobra.Command{
-	Use:   "dock",
-	Short: "manage OpenSDS dock resources",
-	Run:   dockAction,
+var poolCommand = &cobra.Command{
+	Use:   "pool",
+	Short: "manage OpenSDS pool resources",
+	Run:   poolAction,
 }
 
-var dockShowCommand = &cobra.Command{
-	Use:   "show <dock id>",
-	Short: "show information of specified dock",
-	Run:   dockShowAction,
+var poolShowCommand = &cobra.Command{
+	Use:   "show <pool id>",
+	Short: "show information of specified pool",
+	Run:   poolShowAction,
 }
 
-var dockListCommand = &cobra.Command{
+var poolListCommand = &cobra.Command{
 	Use:   "list",
-	Short: "get all dock resources",
-	Run:   dockListAction,
+	Short: "get all pool resources",
+	Run:   poolListAction,
 }
 
 func init() {
-	dockCommand.AddCommand(dockShowCommand)
-	dockCommand.AddCommand(dockListCommand)
+	poolCommand.AddCommand(poolShowCommand)
+	poolCommand.AddCommand(poolListCommand)
 }
 
-func dockAction(cmd *cobra.Command, args []string) {
+func poolAction(cmd *cobra.Command, args []string) {
 	cmd.Usage()
 	os.Exit(1)
 }
 
-func dockShowAction(cmd *cobra.Command, args []string) {
+func poolShowAction(cmd *cobra.Command, args []string) {
 	if len(args) != 1 {
 		fmt.Println("The number of args is not correct!")
 		cmd.Usage()
 		os.Exit(1)
 	}
-	resp, err := client.GetDock(args[0])
+	pols, err := client.GetPool(args[0])
 	if err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
-	keys := KeyList{"Id", "CreatedAt", "UpdatedAt", "Name", "Description", "Endpoint", "DriverName", "Parameters"}
-	PrintDict(resp, keys, FormatterList{})
+	keys := KeyList{"Id", "CreatedAt", "UpdatedAt", "Name", "Description", "Status", "DockId",
+		"AvailabilityZone", "TotalCapacity", "FreeCapacity", "StorageType", "Parameters"}
+	PrintDict(pols, keys, FormatterList{})
 }
 
-func dockListAction(cmd *cobra.Command, args []string) {
+func poolListAction(cmd *cobra.Command, args []string) {
 	if len(args) != 0 {
 		fmt.Println("The number of args is not correct!")
 		cmd.Usage()
 		os.Exit(1)
 	}
-
-	resp, err := client.ListDocks()
+	pols, err := client.ListPools()
 	if err != nil {
 		fmt.Println(err)
 		os.Exit(1)
 	}
-	keys := KeyList{"Id", "Name", "Description", "Endpoint", "DriverName", "Parameters"}
-	PrintList(resp, keys, FormatterList{})
+	keys := KeyList{"Id", "Name", "Description", "Status", "AvailabilityZone", "TotalCapacity", "FreeCapacity"}
+	PrintList(pols, keys, FormatterList{})
 }
-
