+++ pkg/utils/logs/logs.go
@@ -20,6 +20,7 @@ import (
 	"os"
 
 	"github.com/golang/glog"
+	"github.com/opensds/opensds/pkg/utils"
 )
 
 const DefaultLogDir = "/var/log/opensds"
@@ -41,22 +42,11 @@ func (writer GlogWriter) Write(data []byte) (n int, err error) {
 	return len(data), nil
 }
 
-func PathExists(path string) (bool, error) {
-	_, err := os.Stat(path)
-	if err == nil {
-		return true, nil
-	}
-	if os.IsNotExist(err) {
-		return false, nil
-	}
-	return false, err
-}
-
 func InitLogs() {
 	log.SetOutput(GlogWriter{})
 	log.SetFlags(log.LstdFlags | log.Lshortfile)
 	logDir := flag.CommandLine.Lookup("log_dir").Value.String()
-	if exist, _ := PathExists(logDir); !exist {
+	if exist, _ := utils.PathExists(logDir); !exist {
 		os.MkdirAll(logDir, 0755)
 	}
 }
