+++ pkg/utils/utils.go
@@ -24,6 +24,7 @@ import (
 
 	"github.com/opensds/opensds/pkg/model"
 	"github.com/satori/go.uuid"
+	"os"
 )
 
 var S = NewSetter()
@@ -176,4 +177,15 @@ func MergeStringMaps(maps ...map[string]string) map[string]string {
 		}
 	}
 	return out
+}
+
+func PathExists(path string) (bool, error) {
+	_, err := os.Stat(path)
+	if err == nil {
+		return true, nil
+	}
+	if os.IsNotExist(err) {
+		return false, nil
+	}
+	return false, err
 }
